+++ CHANGELOG.md
@@ -7,6 +7,7 @@
  - Replaced all `new_if_supported` or `empty_if_supported` functions with `new` or `empty` and a proper error type.
  - All `new` and `empty` constructors now return an error if the operation is not supported.
  - Removed the deprecated `new_empty` function.
+ - Reworked `TextureCreationError` and removed `TextureMaybeSupportedCreationError`.
 
 ## Version 0.7.1 (2015-07-14)